pyproject.toml
[project]
name = "open-ganss-gs-3104t-pro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.18"
dependencies = [
    "pyusb>=1.3.1",
]
